fixed minor bugs fixed a few odd bugs such as products being added to the previous order and then creating a new order (basket) and then another one with receipts where it was retuning an int instead of a float when calculating total
Added more thematical background and text for the final product for
customer side only, still to apply to admin sections

+++ app/Models/Order.php
@@ -54,14 +54,14 @@ class Order extends Model
     //        PRIVATE FUNCTIONS
     //=====================================
     /**
-     * @return int
+     * @return float
      */
-    private function calcTotal() : int
+    private function calcTotal() : float
     {
-        $total = 0.00;
+        $total = (float) 0.00;
         foreach($this->OrderLines as $ol)
         {
-            $total += $ol->Quantity * $ol->product->Price;
+            $total += (float) $ol->Quantity * $ol->product->Price;
         }
         return $total;
     }
